requirements.txt
django-model-utils>=2.0
https://github.com/matllubos/django-is-core/tarball/NewDjangoSupport#egg=django-is-core-2.8.13
https://github.com/Formulka/django-perms/tarball/0.1.0#egg=django-perms-0.1.0
